Create MaxPool2D_with_Padding_and_Stride.chpl
Please generate the next iteration of this code that does consider the situation where the padding size is greater than the original input size. Also incorporate the ability to handle cases where the input does not evenly divide by the pool size or stride. Please continue to provide error messages and code annotation where appropriate and conclude by providing a file name suggestion.

Here's an updated version of the MaxPool2D class that considers the situations you mentioned. Note that there are many ways to handle the edge cases where the input doesn't evenly divide by the pool size or stride. Here we handle these situations by rounding down, which means we ignore the last few elements of the input if it does not fit an exact number of pool_size or stride.

This file should be saved as MaxPool2D_with_Padding_and_Stride.chpl.

The changes include:

During the forward pass, instead of slicing the padded input with j..#this.pool_size and k..#this.pool_size, we now use j..min(j + this.pool_size - 1, h_pad) and k..min(k + this.pool_size - 1, w_pad). This will slice the input starting from j and k and stop at the smaller of j + this.pool_size - 1 and h_pad, w_pad respectively. This ensures that we don't go out of bounds when j or k are close to the end of the input, and the pool size exceeds the remaining elements.
Similarly, for the backward pass, we use j..min(j + this.pool_size - 1, h*this.stride) and k..min(k + this.pool_size - 1, w*this.stride). This makes sure that the gradient doesn't go out of bounds during back propagation.

Note that we round down the output size by using integer division in the forward pass. This means that if the input size does not divide evenly by the pool size or stride, we simply ignore the last few elements. If you want to handle this differently, you need to adjust the code accordingly.