Create CrossEntropyLossv2.chpl
Here is a revised version of CrossEntropyLoss.chpl with added checks for numerical stability and support for batch inputs:

In this version, we added a small constant epsilon to the prediction probabilities before taking the log to prevent taking the log of a very small number. The value of epsilon can be customized when initializing the CrossEntropyLoss object. By default, it's set to 1e-8.

This file can be saved as CrossEntropyLoss.chpl.

It's important to note that this implementation assumes that the input tensors are 2-dimensional, where the first dimension represents the batch size (i.e., the number of instances), and the second dimension represents the number of classes. The prediction probabilities should be normalized along the second dimension before passing to this function.