Create Utils.chpl
Creating utility functions in Chapel depends highly on the specifics of your project and its requirements. However, here is an example of what a Utils.chpl file might look like for a machine learning project in Chapel:

In this Utils module, we have defined several utility procedures that can be used in a machine learning project:

shuffle: Shuffles two arrays in-place using the given seed. This could be useful for shuffling your dataset before each epoch during training.
trainTestSplit: Splits an array into a training set and a test set with the given ratio. This could be used to create a validation set from your training data.
accuracy: Computes the accuracy of the predicted labels against the true labels. This could be used to evaluate the performance of your model after each epoch.
standardize: Standardizes an array using mean normalization. This could be used for preprocessing your data before feeding it to the model.
These are just a few examples of utility functions that could be useful in a machine learning project. Depending on your project's requirements, you may need to define additional utility functions.